### 2016-09-26 #81 @whedon
[PRE REVIEW]: Correlation Trait Loci (CTL) mapping: phenotype network inference subject to genotype
**Submitting author:** @DannyArends (<a href="http://orcid.org/0000-0001-8738-0162">Danny Arends</a>)
**Repository:** <a href="https://github.com/DannyArends/CTLmapping" target ="_blank">https://github.com/DannyArends/CTLmapping</a>
**Version:** v1.0.0-joss-1
