### 2016-09-27 e2b9c52
Updating the DOI badge to make sure citations should go to v1.0.0 release of Zenodo
