### 2016-09-28 58d72c5
Adding a reference to my thesis in Algorithm.md
