### 2016-09-28 778daa9
Slightly updated the text of algorithm.md
