+++ doc/ALGORITHM.md
@@ -12,5 +12,5 @@ genetic wiring of the classical phenotypes and identify key players in the
 genetic / protein network underlying classical phenotypes using QTL and CTL 
 information.
 
-For more information on how-to calculate CTLs on N-genotypes, please read 
-Chapter 4 of [my Thesis](http://hdl.handle.net/11370/29cf3cc5-6596-4d5a-a87c-490b31676a96)
+For more information on CTL and their interaction with QTLs, or how-to calculate 
+CTLs on N-genotypes, please see Chapter 4 of [my Doctoral Thesis](http://hdl.handle.net/11370/29cf3cc5-6596-4d5a-a87c-490b31676a96)
