### 2016-10-07 #87 @whedon
[REVIEW]: Correlation Trait Loci (CTL) mapping: phenotype network inference subject to genotype
**Submitting author:** @DannyArends (<a href="http://orcid.org/0000-0001-8738-0162">Danny Arends</a>)
**Repository:** <a href="https://github.com/DannyArends/CTLmapping" target ="_blank">https://github.com/DannyArends/CTLmapping</a>
**Version:** v1.0.0-joss-2
**Archive:** <a href="https://dx.doi.org/10.5281/zenodo.163510" target="_blank">https://dx.doi.org/10.5281/zenodo.163510</a>

- [x] **Version:** Does the release version given match the GitHub release (v1.0.0-joss-1)?
