### 2016-10-07 #87 @arfon
@byandell this is where you should report back on your review. Please update the checklist on this issue as you proceed.
