### 2016-10-08 #87 @pjotrp
The paper URL is not complete. You can find it at https://github.com/DannyArends/CTLmapping/blob/master/doc/joss/2016/paper.md
