### 2016-10-08 #87 @arfon
> The paper URL is not complete. You can find it at https://github.com/DannyArends/CTLmapping/blob/master/doc/joss/2016/paper.md

Thanks for the reminder @pjotrp. @byandell - I've compiled the paper for your review: [10.21105.joss.00087.pdf](https://github.com/openjournals/joss-reviews/files/517691/10.21105.joss.00087.pdf)
