### 2016-10-09 50de7d9
Removed a double affiliation (Pjotr Prins) in the JOSS paper
