### 2016-10-14 #87 @byandell
I updated the comment by arfon, which said I should put my review there. I took this literally and edited that entry to include my review. I attache below as well.
Brian

Review of Danny Arends

Would be very helpful to plot St7 against Il18r1 using plot symbols for genotype at strongest CTL signal on chr 2 at ~15Mbp. Similarly do this for Mtvr2 against C1qtnf5 on chr 19 proximal end with CTL genotype. If first case, expect to see a change in direction of association.

What is basis for "rule of thumb"? Does this involve some genome-wide correction or threshold calculation?

Here is some related work that probably should be cited.

Relate CTL to ANCOVA which is the underlying tool to determine change in correlation. Usual use of ANCOVA in QTL is to adjust for covariates in searching for QTL (see Broman and Sen 2009 book).

CTL differs fundamentally from mediation (Chick et al. Gygi 2016 doi:10.1038/nature18270) where the goal is to use covariates tied to genomic position (e.g. mRNA expression) to look at change in strength of QTL signal.

For MAGIC, please cite de Koning and McIntyre LM (2014) doi:10.1534/g3.114.013680
