### 2016-10-14 #87 @pjotrp
Thanks @byandell! We'll amend the paper to include your suggestions.
