### 2016-10-24 f6a9dc3
Adding the citation for the MAGIC population (DeKoning:2014)
