+++ doc/joss/2016/paper.bib
@@ -17,6 +17,22 @@
   year = {2010}
 }
 
+@article{DeKoning:2014,
+  author = {de Koning, D.J. and McIntyre, Lauren M},
+  doi = {10.1534/g3.114.013680},
+  issn = {2160-1836},
+  journal = {G3{\&}{\#}58; Genes|Genomes|Genetics},
+  month = {sep},
+  number = {9},
+  pages = {1567--1568},
+  publisher = {Genetics Society of America},
+  title = {{GENETICS and G3: Community-Driven Science, Community-Driven Journals}},
+  url = {http://www.ncbi.nlm.nih.gov/pmc/articles/PMC4169148/ http://g3journal.org/cgi/doi/10.1534/g3.114.013680},
+  volume = {4},
+  year = {2014}
+}
+
+
 @article{Cytoscape:2003,
   abstract = {Cytoscape is an open source software project for integrating biomolecular interaction networks with high-throughput expression data and other molecular states into a unified conceptual framework. Although applicable to any system of molecular components and interactions, Cytoscape is most powerful when used in conjunction with large databases of protein-protein, protein-DNA, and genetic interactions that are increasingly available for humans and model organisms. Cytoscape's software Core provides basic functionality to layout and query the network; to visually integrate the network with expression profiles, phenotypes, and other molecular states; and to link the network to databases of functional annotations. The Core is extensible through a straightforward plug-in architecture, allowing rapid development of additional computational analyses and features. Several case studies of Cytoscape plug-ins are surveyed, including a search for interaction pathways correlating with changes in gene expression, a study of protein complexes involved in cellular recovery to DNA damage, inference of a combined physical/functional interaction network for Halobacterium, and an interface to detailed stochastic/kinetic gene regulatory models.},
   author = {Shannon, P and Markiel, A and Ozier, O and Baliga, N S and Wang, J T and Ramage, D and Amin, N and Schwikowski, B and Ideker, T},
