### 2016-10-24 113e27d
Adding a sentence about how the rule of thumb was determined (simulations)
