### 2016-10-24 750e545
Updating the text, to add references to the ANCOVA and mediation methodolody, as suggested by B. Yandell
