+++ doc/joss/2016/paper.bib
@@ -32,6 +32,22 @@
   year = {2014}
 }
 
+@article{Chick2016,
+abstract = {Genetic variation modulates protein expression through both transcriptional and post-transcriptional mechanisms. To characterize the consequences of natural genetic diversity on the proteome, here we combine a multiplexed, mass spectrometry-based method for protein quantification with an emerging outbred mouse model containing extensive genetic variation from eight inbred founder strains. By measuring genome-wide transcript and protein expression in livers from 192 Diversity outbred mice, we identify 2,866 protein quantitative trait loci (pQTL) with twice as many local as distant genetic variants. These data support distinct transcriptional and post-transcriptional models underlying the observed pQTL effects. Using a sensitive approach to mediation analysis, we often identified a second protein or transcript as the causal mediator of distant pQTL. Our analysis reveals an extensive network of direct protein–protein interactions. Finally, we show that local genotype can provide accurate predictions of protein abundance in an independent cohort of collaborative cross mice.},
+author = {Chick, Joel M and Munger, Steven C and Simecek, Petr and Huttlin, Edward L and Choi, Kwangbom and Gatti, Daniel M and Raghupathy, Narayanan and Svenson, Karen L and Churchill, Gary A and Gygi, Steven P},
+doi = {10.1038/nature18270},
+issn = {0028-0836},
+journal = {Nature},
+month = {jun},
+number = {7608},
+pages = {500--505},
+publisher = {Nature Publishing Group, a division of Macmillan Publishers Limited. All Rights Reserved.},
+title = {{Defining the consequences of genetic variation on a proteome-wide scale}},
+url = {http://dx.doi.org/10.1038/nature18270 http://10.1038/nature18270 http://www.nature.com/nature/journal/v534/n7608/abs/nature18270.html{\#}supplementary-information http://www.nature.com/doifinder/10.1038/nature18270},
+volume = {534},
+year = {2016}
+}
+
 
 @article{Cytoscape:2003,
   abstract = {Cytoscape is an open source software project for integrating biomolecular interaction networks with high-throughput expression data and other molecular states into a unified conceptual framework. Although applicable to any system of molecular components and interactions, Cytoscape is most powerful when used in conjunction with large databases of protein-protein, protein-DNA, and genetic interactions that are increasingly available for humans and model organisms. Cytoscape's software Core provides basic functionality to layout and query the network; to visually integrate the network with expression profiles, phenotypes, and other molecular states; and to link the network to databases of functional annotations. The Core is extensible through a straightforward plug-in architecture, allowing rapid development of additional computational analyses and features. Several case studies of Cytoscape plug-ins are surveyed, including a search for interaction pathways correlating with changes in gene expression, a study of protein complexes involved in cellular recovery to DNA damage, inference of a combined physical/functional interaction network for Halobacterium, and an interface to detailed stochastic/kinetic gene regulatory models.},
