### 2016-10-24 f36646d
JOSS paper: some edits
