### 2016-10-24 85cab4e
Fixing the missing : in Chick:2016 citation, and removing a space
