+++ doc/joss/2016/paper.bib
@@ -32,20 +32,20 @@
   year = {2014}
 }
 
-@article{Chick2016,
-abstract = {Genetic variation modulates protein expression through both transcriptional and post-transcriptional mechanisms. To characterize the consequences of natural genetic diversity on the proteome, here we combine a multiplexed, mass spectrometry-based method for protein quantification with an emerging outbred mouse model containing extensive genetic variation from eight inbred founder strains. By measuring genome-wide transcript and protein expression in livers from 192 Diversity outbred mice, we identify 2,866 protein quantitative trait loci (pQTL) with twice as many local as distant genetic variants. These data support distinct transcriptional and post-transcriptional models underlying the observed pQTL effects. Using a sensitive approach to mediation analysis, we often identified a second protein or transcript as the causal mediator of distant pQTL. Our analysis reveals an extensive network of direct protein–protein interactions. Finally, we show that local genotype can provide accurate predictions of protein abundance in an independent cohort of collaborative cross mice.},
-author = {Chick, Joel M and Munger, Steven C and Simecek, Petr and Huttlin, Edward L and Choi, Kwangbom and Gatti, Daniel M and Raghupathy, Narayanan and Svenson, Karen L and Churchill, Gary A and Gygi, Steven P},
-doi = {10.1038/nature18270},
-issn = {0028-0836},
-journal = {Nature},
-month = {jun},
-number = {7608},
-pages = {500--505},
-publisher = {Nature Publishing Group, a division of Macmillan Publishers Limited. All Rights Reserved.},
-title = {{Defining the consequences of genetic variation on a proteome-wide scale}},
-url = {http://dx.doi.org/10.1038/nature18270 http://10.1038/nature18270 http://www.nature.com/nature/journal/v534/n7608/abs/nature18270.html{\#}supplementary-information http://www.nature.com/doifinder/10.1038/nature18270},
-volume = {534},
-year = {2016}
+@article{Chick:2016,
+  abstract = {Genetic variation modulates protein expression through both transcriptional and post-transcriptional mechanisms. To characterize the consequences of natural genetic diversity on the proteome, here we combine a multiplexed, mass spectrometry-based method for protein quantification with an emerging outbred mouse model containing extensive genetic variation from eight inbred founder strains. By measuring genome-wide transcript and protein expression in livers from 192 Diversity outbred mice, we identify 2,866 protein quantitative trait loci (pQTL) with twice as many local as distant genetic variants. These data support distinct transcriptional and post-transcriptional models underlying the observed pQTL effects. Using a sensitive approach to mediation analysis, we often identified a second protein or transcript as the causal mediator of distant pQTL. Our analysis reveals an extensive network of direct protein–protein interactions. Finally, we show that local genotype can provide accurate predictions of protein abundance in an independent cohort of collaborative cross mice.},
+  author = {Chick, Joel M and Munger, Steven C and Simecek, Petr and Huttlin, Edward L and Choi, Kwangbom and Gatti, Daniel M and Raghupathy, Narayanan and Svenson, Karen L and Churchill, Gary A and Gygi, Steven P},
+  doi = {10.1038/nature18270},
+  issn = {0028-0836},
+  journal = {Nature},
+  month = {jun},
+  number = {7608},
+  pages = {500--505},
+  publisher = {Nature Publishing Group, a division of Macmillan Publishers Limited. All Rights Reserved.},
+  title = {{Defining the consequences of genetic variation on a proteome-wide scale}},
+  url = {http://dx.doi.org/10.1038/nature18270 http://10.1038/nature18270 http://www.nature.com/nature/journal/v534/n7608/abs/nature18270.html{\#}supplementary-information http://www.nature.com/doifinder/10.1038/nature18270},
+  volume = {534},
+  year = {2016}
 }
 
 
