### 2016-10-26 #87 @pjotrp
@byandell we modified the paper following your suggestions: https://github.com/DannyArends/CTLmapping/blob/master/doc/joss/2016/paper.md. @arfon as the software ticks all the boxes I believe we can accept the paper and go live.
