### 2016-10-27 #87 @DannyArends
@arfon I created a new release (v1.0.0-joss-2) with Zenodo, the DOI of the release is: 10.5281/zenodo.163510
