### 2016-10-27 #87 @byandell
I am ok with paper as is. I would have liked to see scatterplot of variables with QTL alleles as plot symbols, but not necessary. I submitted very minor typo as issue just now. Brian

Get Outlook for iOS<https://aka.ms/o0ukef>
