### 2016-10-30 #87 @arfon
@byandell many thanks for your review!

@DannyArends your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00087 🚀 🎉 💥
